requirements.txt
# ===== CORE WEB FRAMEWORK =====
flask==3.1.1
werkzeug==3.1.0
jinja2==3.1.4

# ===== OPENAI INTEGRATION =====
openai==1.107.1
python-dotenv==1.0.1

# ===== DATA PROCESSING (Python 3.13 Compatible) =====
numpy==2.1.1
pandas==2.2.3
scikit-learn==1.5.2

# ===== VISUALIZATION & CHARTS =====
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# ===== PDF REPORT GENERATION =====
reportlab==4.2.5

# ===== WEB UTILITIES =====
flask-cors==5.0.0
requests==2.32.3

# ===== DATA SCIENCE & ML =====
scipy==1.14.1
imbalanced-learn==0.12.4

# ===== OPTIONAL API FRAMEWORK =====
fastapi==0.115.0
uvicorn==0.30.6

# ===== DEVELOPMENT TOOLS =====
pytest==8.3.3
black==24.8.0
flake8==7.1.1
